Add a first test case for instant answer
